# kommonitor-datamanagement-api-client/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: christian.danowski-buhren@hs-bochum.de
  description: "Definition einer Datenzugriffs-API, die den Zugriff auf die KomMonitor-Datenhaltungsschicht\
    \ kapselt."
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: KomMonitor Data Access API
  version: 0.0.1
servers:
- url: //localhost:8085/
...
paths:
...
  /spatial-units:
    get:
      deprecated: false
      description: retrieve information about available features of different spatial
        units/levels
      operationId: getSpatialUnits
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/SpatialUnitOverviewType"
                type: array
          description: OK
        "400":
          content: {}
          description: Invalid status value
        "401":
          content: {}
          description: API key is missing or invalid
        "403":
          content: {}
          description: Forbidden
        "404":
          content: {}
          description: Not Found
      security:
      - kommonitor-data-access_oauth: []
      summary: retrieve information about available features of different spatial
        units/levels
      tags:
      - spatial-units
      x-accepts:
      - application/json
    post:
      deprecated: false
      description: Add/Register a spatial unit for a certain period of time
      operationId: addSpatialUnitAsBody
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpatialUnitPOSTInputType"
        description: feature data
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpatialUnitOverviewType"
          description: OK
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpatialUnitOverviewType"
          description: Created
        "401":
          content: {}
          description: API key is missing or invalid
        "403":
          content: {}
          description: Forbidden
        "404":
          content: {}
          description: Not Found
        "405":
          content: {}
          description: Invalid input
      security:
      - kommonitor-data-access_oauth: []
      summary: Add a new spatial-unit
      tags:
      - spatial-units
      x-codegen-request-body-name: featureData
      x-content-type: application/json
      x-accepts:
      - application/json
...
components:
  schemas:
    CommonMetadataType:
      example:
        literature: literature
        note: note
        updateInterval: ARBITRARY
        sridEPSG: 0.0
        datasource: datasource
        contact: contact
        lastUpdate: 2000-01-23
        description: description
        databasis: databasis
      properties:
        contact:
          description: contact details where additional information can be achieved
          type: string
        databasis:
          description: information about data used as a basis to generate the dataset
          type: string
        datasource:
          description: information about the origin/source of the dataset
          type: string
        description:
          description: description of the dataset
          type: string
        lastUpdate:
          description: a timestamp representing the lastUpdate according to ISO 8601
            (e.g. 2018-01-30)
          format: date
          type: string
        literature:
          description: an optional hint to literature about the dataset (e.g. URL
            or book/article name)
          type: string
        note:
          description: an optional note with background information about the dataset
          type: string
        sridEPSG:
          description: the coordinate reference system of the dataset as EPSG code
          example: 0.0
          type: number
        updateInterval:
          enum:
          - ARBITRARY
          - MONTHLY
          - QUARTERLY
          - HALF_YEARLY
          - YEARLY
          - DAILY
          - WEEKLY
          type: string
      required:
      - contact
      - datasource
      - description
      - sridEPSG
      - updateInterval
      title: CommonMetadataType
      type: object
...
    PeriodOfValidityType:
      description: definition of the period of validity of a certain dataset
      example:
        endDate: 2000-01-23
        startDate: 2000-01-23
      properties:
        endDate:
          description: "an optional timestamp representing the ending date according\
            \ to ISO 8601 (e.g. 2018-01-30). The parameter can be omitted, if the\
            \ end date is unknown."
          format: date
          type: string
        startDate:
          description: a timestamp representing the starting date according to ISO
            8601 (e.g. 2018-01-30)
          format: date
          type: string
      required:
      - startDate
      title: PeriodOfValidityType
      type: object
...
    SpatialUnitOverviewType:
      example:
        metadata:
          literature: literature
          note: note
          updateInterval: ARBITRARY
          sridEPSG: 0.0
          datasource: datasource
          contact: contact
          lastUpdate: 2000-01-23
          description: description
          databasis: databasis
        outlineDashArrayString: outlineDashArrayString
        wfsUrl: wfsUrl
        isOutlineLayer: true
        nextLowerHierarchyLevel: nextLowerHierarchyLevel
        userPermissions:
        - creator
        - creator
        ownerId: ownerId
        availablePeriodsOfValidity:
        - endDate: 2000-01-23
          startDate: 2000-01-23
        - endDate: 2000-01-23
          startDate: 2000-01-23
        spatialUnitId: spatialUnitId
        wmsUrl: wmsUrl
        outlineWidth: 0.8008281904610115
        permissions:
        - permissions
        - permissions
        outlineColor: outlineColor
        isPublic: true
        spatialUnitLevel: spatialUnitLevel
        nextUpperHierarchyLevel: nextUpperHierarchyLevel
      properties:
        permissions:
          description: list of permissions on this entity
          items:
            type: string
          type: array
        availablePeriodsOfValidity:
          items:
            $ref: "#/components/schemas/PeriodOfValidityType"
          type: array
        isPublic:
          description: flag whether the resource is publicly accessible
          type: boolean
        metadata:
          $ref: "#/components/schemas/CommonMetadataType"
        nextLowerHierarchyLevel:
          description: the identifier/name of the spatial unit level that contains
            the features of the nearest lower hierarchy level
          type: string
        nextUpperHierarchyLevel:
          description: the identifier/name of the spatial unit level that contains
            the features of the nearest upper hierarchy level
          type: string
        spatialUnitId:
          description: the unique identifier of the spatial unit level the features
            apply to
          type: string
        spatialUnitLevel:
          description: the name of the spatial unit level the features apply to
          type: string
        userPermissions:
          description: list of permissions that are effective on this dataset for
            the current user
          items:
            $ref: "#/components/schemas/PermissionLevelType"
          type: array
        wfsUrl:
          description: the URL of a running WFS instance serving the spatial features
            of the associated dataset
          type: string
        wmsUrl:
          description: the URL of a running WMS instance serving the spatial features
            of the associated dataset
          type: string
        isOutlineLayer:
          description: "if true, then KomMonitor web client map application will offer\
            \ this spatial unit as outline layer in legend control"
          type: boolean
        outlineColor:
          description: outline color for this layer as hex code
          type: string
        outlineWidth:
          description: outline width as stroke width for outline geometry
          format: int32
          type: number
        outlineDashArrayString:
          description: "string of line stroke dash array for lines of interest (e.g.\
            \ 20,20; see https://developer.mozilla.org/de/docs/Web/SVG/Attribute/stroke-dasharray)"
          type: string
        ownerId:
          description: identifier of the owning group
          type: string
      required:
      - isPublic
      - metadata
      - nextLowerHierarchyLevel
      - nextUpperHierarchyLevel
      - permissions
      - spatialUnitId
      - spatialUnitLevel
      - userPermissions
      - wfsUrl
      - wmsUrl
      title: SpatialUnitOverviewType
      type: object
...
    SpatialUnitPOSTInputType:
      example:
        metadata:
          literature: literature
          note: note
          updateInterval: ARBITRARY
          sridEPSG: 0.0
          datasource: datasource
          contact: contact
          lastUpdate: 2000-01-23
          description: description
          databasis: databasis
        outlineDashArrayString: outlineDashArrayString
        isOutlineLayer: false
        nextLowerHierarchyLevel: nextLowerHierarchyLevel
        ownerId: ownerId
        geoJsonString: geoJsonString
        outlineWidth: 0.8008281904610115
        jsonSchema: jsonSchema
        permissions:
        - permissions
        - permissions
        outlineColor: outlineColor
        isPublic: true
        periodOfValidity:
          endDate: 2000-01-23
          startDate: 2000-01-23
        spatialUnitLevel: spatialUnitLevel
        nextUpperHierarchyLevel: nextUpperHierarchyLevel
      properties:
        permissions:
          description: list of permissions on this entity
          items:
            type: string
          type: array
        geoJsonString:
          description: a valid GeoJSON string containing the features consisting of
            a geometry and a unique identifier as property 'uuid'
          type: string
        jsonSchema:
          description: a JSON schema as string that defines the data model for this
            dataset. It can be used to validate the geoJsonString property.
          type: string
        metadata:
          $ref: "#/components/schemas/CommonMetadataType"
        nextLowerHierarchyLevel:
          description: the identifier/name of the spatial unit level that contains
            the features of the nearest lower hierarchy level
          type: string
        nextUpperHierarchyLevel:
          description: the identifier/name of the spatial unit level that contains
            the features of the nearest upper hierarchy level
          type: string
        periodOfValidity:
          $ref: "#/components/schemas/PeriodOfValidityType"
        spatialUnitLevel:
          description: the name and identifier of the spatial unit level the features
            apply to
          type: string
        isOutlineLayer:
          default: false
          description: "if true, then KomMonitor web client map application will offer\
            \ this spatial unit as outline layer in legend control"
          type: boolean
        outlineColor:
          description: outline color for this layer as hex code
          type: string
        outlineWidth:
          description: outline width as stroke width for outline geometry
          format: int32
          type: number
        outlineDashArrayString:
          description: "string of line stroke dash array for lines of interest (e.g.\
            \ 20,20; see https://developer.mozilla.org/de/docs/Web/SVG/Attribute/stroke-dasharray)"
          type: string
        ownerId:
          description: identifier of the owning group
          type: string
        isPublic:
          description: flag whether the resource is publicly accessible
          type: boolean
      required:
      - geoJsonString
      - isPublic
      - metadata
      - periodOfValidity
      - permissions
      - spatialUnitLevel
      title: SpatialUnitPOSTInputType
      type: object
...
    PermissionLevelType:
      description: Permission Levels for CRUD operations.
      enum:
      - creator
      - editor
      - viewer
      type: string
...
  securitySchemes:
    kommonitor-data-access_oauth:
      flows:
        authorizationCode:
          authorizationUrl: http://keycloak:8080/realms/kommonitor/protocol/openid-connect/auth
          scopes: {}
          tokenUrl: http://keycloak:8080/realms/kommonitor/protocol/openid-connect/token
      type: oauth2
...